+++ test/ITF1788_tests/libieeep1788_tests_num.jl
@@ -31,71 +31,71 @@ setprecision(Interval, Float64)
 # setrounding(Interval, :narrow)
 
 facts("minimal_inf_test") do
-    @fact infimum(∅) --> Inf
-    @fact infimum(Interval(-Inf, Inf)) --> -Inf
-    @fact infimum(Interval(1.0, 2.0)) --> 1.0
-    @fact infimum(Interval(-3.0, -2.0)) --> -3.0
-    @fact infimum(Interval(-Inf, 2.0)) --> -Inf
-    @fact infimum(Interval(-Inf, 0.0)) --> -Inf
-    @fact infimum(Interval(-Inf, -0.0)) --> -Inf
-    @fact infimum(Interval(-2.0, Inf)) --> -2.0
-    @fact infimum(Interval(0.0, Inf)) --> -0.0
-    @fact infimum(Interval(-0.0, Inf)) --> -0.0
-    @fact infimum(Interval(-0.0, 0.0)) --> -0.0
-    @fact infimum(Interval(0.0, -0.0)) --> -0.0
-    @fact infimum(Interval(0.0, 0.0)) --> -0.0
-    @fact infimum(Interval(-0.0, -0.0)) --> -0.0
+    @fact inf(∅) --> Inf
+    @fact inf(Interval(-Inf, Inf)) --> -Inf
+    @fact inf(Interval(1.0, 2.0)) --> 1.0
+    @fact inf(Interval(-3.0, -2.0)) --> -3.0
+    @fact inf(Interval(-Inf, 2.0)) --> -Inf
+    @fact inf(Interval(-Inf, 0.0)) --> -Inf
+    @fact inf(Interval(-Inf, -0.0)) --> -Inf
+    @fact inf(Interval(-2.0, Inf)) --> -2.0
+    @fact inf(Interval(0.0, Inf)) --> -0.0
+    @fact inf(Interval(-0.0, Inf)) --> -0.0
+    @fact inf(Interval(-0.0, 0.0)) --> -0.0
+    @fact inf(Interval(0.0, -0.0)) --> -0.0
+    @fact inf(Interval(0.0, 0.0)) --> -0.0
+    @fact inf(Interval(-0.0, -0.0)) --> -0.0
 end
 
 facts("minimal_inf_dec_test") do
-    @fact infimum(DecoratedInterval(∅, trv)) --> Inf
-    @fact infimum(DecoratedInterval(Interval(-Inf, Inf), def)) --> -Inf
-    @fact infimum(DecoratedInterval(Interval(1.0, 2.0), com)) --> 1.0
-    @fact infimum(DecoratedInterval(Interval(-3.0, -2.0), trv)) --> -3.0
-    @fact infimum(DecoratedInterval(Interval(-Inf, 2.0), dac)) --> -Inf
-    @fact infimum(DecoratedInterval(Interval(-Inf, 0.0), def)) --> -Inf
-    @fact infimum(DecoratedInterval(Interval(-Inf, -0.0), trv)) --> -Inf
-    @fact infimum(DecoratedInterval(Interval(-2.0, Inf), trv)) --> -2.0
-    @fact infimum(DecoratedInterval(Interval(0.0, Inf), def)) --> -0.0
-    @fact infimum(DecoratedInterval(Interval(-0.0, Inf), trv)) --> -0.0
-    @fact infimum(DecoratedInterval(Interval(-0.0, 0.0), dac)) --> -0.0
-    @fact infimum(DecoratedInterval(Interval(0.0, -0.0), trv)) --> -0.0
-    @fact infimum(DecoratedInterval(Interval(0.0, 0.0), trv)) --> -0.0
-    @fact infimum(DecoratedInterval(Interval(-0.0, -0.0), trv)) --> -0.0
+    @fact inf(DecoratedInterval(∅, trv)) --> Inf
+    @fact inf(DecoratedInterval(Interval(-Inf, Inf), def)) --> -Inf
+    @fact inf(DecoratedInterval(Interval(1.0, 2.0), com)) --> 1.0
+    @fact inf(DecoratedInterval(Interval(-3.0, -2.0), trv)) --> -3.0
+    @fact inf(DecoratedInterval(Interval(-Inf, 2.0), dac)) --> -Inf
+    @fact inf(DecoratedInterval(Interval(-Inf, 0.0), def)) --> -Inf
+    @fact inf(DecoratedInterval(Interval(-Inf, -0.0), trv)) --> -Inf
+    @fact inf(DecoratedInterval(Interval(-2.0, Inf), trv)) --> -2.0
+    @fact inf(DecoratedInterval(Interval(0.0, Inf), def)) --> -0.0
+    @fact inf(DecoratedInterval(Interval(-0.0, Inf), trv)) --> -0.0
+    @fact inf(DecoratedInterval(Interval(-0.0, 0.0), dac)) --> -0.0
+    @fact inf(DecoratedInterval(Interval(0.0, -0.0), trv)) --> -0.0
+    @fact inf(DecoratedInterval(Interval(0.0, 0.0), trv)) --> -0.0
+    @fact inf(DecoratedInterval(Interval(-0.0, -0.0), trv)) --> -0.0
 end
 
 facts("minimal_sup_test") do
-    @fact supremum(∅) --> -Inf
-    @fact supremum(Interval(-Inf, Inf)) --> Inf
-    @fact supremum(Interval(1.0, 2.0)) --> 2.0
-    @fact supremum(Interval(-3.0, -2.0)) --> -2.0
-    @fact supremum(Interval(-Inf, 2.0)) --> 2.0
-    @fact supremum(Interval(-Inf, 0.0)) --> 0.0
-    @fact supremum(Interval(-Inf, -0.0)) --> 0.0
-    @fact supremum(Interval(-2.0, Inf)) --> Inf
-    @fact supremum(Interval(0.0, Inf)) --> Inf
-    @fact supremum(Interval(-0.0, Inf)) --> Inf
-    @fact supremum(Interval(-0.0, 0.0)) --> 0.0
-    @fact supremum(Interval(0.0, -0.0)) --> 0.0
-    @fact supremum(Interval(0.0, 0.0)) --> 0.0
-    @fact supremum(Interval(-0.0, -0.0)) --> 0.0
+    @fact sup(∅) --> -Inf
+    @fact sup(Interval(-Inf, Inf)) --> Inf
+    @fact sup(Interval(1.0, 2.0)) --> 2.0
+    @fact sup(Interval(-3.0, -2.0)) --> -2.0
+    @fact sup(Interval(-Inf, 2.0)) --> 2.0
+    @fact sup(Interval(-Inf, 0.0)) --> 0.0
+    @fact sup(Interval(-Inf, -0.0)) --> 0.0
+    @fact sup(Interval(-2.0, Inf)) --> Inf
+    @fact sup(Interval(0.0, Inf)) --> Inf
+    @fact sup(Interval(-0.0, Inf)) --> Inf
+    @fact sup(Interval(-0.0, 0.0)) --> 0.0
+    @fact sup(Interval(0.0, -0.0)) --> 0.0
+    @fact sup(Interval(0.0, 0.0)) --> 0.0
+    @fact sup(Interval(-0.0, -0.0)) --> 0.0
 end
 
 facts("minimal_sup_dec_test") do
-    @fact supremum(DecoratedInterval(∅, trv)) --> -Inf
-    @fact supremum(DecoratedInterval(Interval(-Inf, Inf), def)) --> Inf
-    @fact supremum(DecoratedInterval(Interval(1.0, 2.0), com)) --> 2.0
-    @fact supremum(DecoratedInterval(Interval(-3.0, -2.0), trv)) --> -2.0
-    @fact supremum(DecoratedInterval(Interval(-Inf, 2.0), dac)) --> 2.0
-    @fact supremum(DecoratedInterval(Interval(-Inf, 0.0), def)) --> 0.0
-    @fact supremum(DecoratedInterval(Interval(-Inf, -0.0), trv)) --> 0.0
-    @fact supremum(DecoratedInterval(Interval(-2.0, Inf), trv)) --> Inf
-    @fact supremum(DecoratedInterval(Interval(0.0, Inf), def)) --> Inf
-    @fact supremum(DecoratedInterval(Interval(-0.0, Inf), trv)) --> Inf
-    @fact supremum(DecoratedInterval(Interval(-0.0, 0.0), dac)) --> +0.0
-    @fact supremum(DecoratedInterval(Interval(0.0, -0.0), trv)) --> +0.0
-    @fact supremum(DecoratedInterval(Interval(0.0, 0.0), trv)) --> +0.0
-    @fact supremum(DecoratedInterval(Interval(-0.0, -0.0), trv)) --> +0.0
+    @fact sup(DecoratedInterval(∅, trv)) --> -Inf
+    @fact sup(DecoratedInterval(Interval(-Inf, Inf), def)) --> Inf
+    @fact sup(DecoratedInterval(Interval(1.0, 2.0), com)) --> 2.0
+    @fact sup(DecoratedInterval(Interval(-3.0, -2.0), trv)) --> -2.0
+    @fact sup(DecoratedInterval(Interval(-Inf, 2.0), dac)) --> 2.0
+    @fact sup(DecoratedInterval(Interval(-Inf, 0.0), def)) --> 0.0
+    @fact sup(DecoratedInterval(Interval(-Inf, -0.0), trv)) --> 0.0
+    @fact sup(DecoratedInterval(Interval(-2.0, Inf), trv)) --> Inf
+    @fact sup(DecoratedInterval(Interval(0.0, Inf), def)) --> Inf
+    @fact sup(DecoratedInterval(Interval(-0.0, Inf), trv)) --> Inf
+    @fact sup(DecoratedInterval(Interval(-0.0, 0.0), dac)) --> +0.0
+    @fact sup(DecoratedInterval(Interval(0.0, -0.0), trv)) --> +0.0
+    @fact sup(DecoratedInterval(Interval(0.0, 0.0), trv)) --> +0.0
+    @fact sup(DecoratedInterval(Interval(-0.0, -0.0), trv)) --> +0.0
 end
 
 facts("minimal_mid_test") do